% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcpip.R
\name{gcpip_ranges}
\alias{gcpip_ranges}
\title{Fetch Google Cloud IP address ranges}
\usage{
gcpip_ranges(service = NULL, scope = NULL)
}
\arguments{
\item{service}{Optional filter by service name (e.g. "Google Cloud").
Case-insensitive.}

\item{scope}{Optional filter by scope/region (e.g. "us-east1", "europe-west1")}
}
\value{
tibble: ip_prefix, service, scope
}
\description{
Returns the current list of Google Cloud IP address ranges, optionally
filtered by service and/or scope (region).
}
