% Generated by build script
\name{gcpip_ranges}
\alias{gcpip_ranges}
\title{Fetch Google Cloud IP address ranges}
\usage{
gcpip_ranges()
}
\arguments{
\item{service}{Optional filter by service name (e.g. "Google Cloud").}
\item{scope}{Optional filter by scope/region (e.g. "us-east1", "europe-west1")}
}
\value{tibble: ip_prefix, service, scope}
\description{Fetch Google Cloud IP address ranges}
